+++ src/NALib/NAEnvironment/Windows/WINAPI/C/UI/NAWindowII.h
@@ -376,7 +376,7 @@ NA_HDEF NARect naGetWindowAbsoluteInnerRect(NACoreUIElement* window){
   screenrect = naGetMainScreenRect();
 
   rect.pos.x = testpoint.x;
-  rect.pos.y = screenrect.size.height - ((double)testpoint.y + (contentrect.bottom - contentrect.top));
+  rect.pos.y = (double)screenrect.size.height - ((double)testpoint.y + ((double)contentrect.bottom - (double)contentrect.top));
   rect.size.width = (double)contentrect.right - (double)contentrect.left;
   rect.size.height = (double)contentrect.bottom - (double)contentrect.top;
   return rect;
